pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "os-tester"
version = "0.1.0"
authors = [
  { name="Fabian Sauter", email="fabian.sauter+pip@apsensing.com" }
]
maintainers = [
  { name="Fabian Sauter", email="fabian.sauter+pip@apsensing.com" }
]
description = "A Python pip package to automate testing of whole operating systems with an image recognition based approach and libvirt (qemu). Inspired by openQA."
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English"
]
dependencies = ["scikit-image", "opencv-python", "libvirt-python", "PyYAML", "numpy", "matplotlib"]

[project.urls]
"Homepage" = "https://github.com/AP-Sensing/os-tester"
"Bug Tracker" = "https://github.com/AP-Sensing/os-tester/issues"
